requirements.txt
asgiref==3.8.1
Django==5.0.7
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==26.0.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
